Flattened new command
